requirements.txt
asgiref==3.7.2
attrs==23.2.0
coverage==7.4.3
Django==5.0.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
django-js-asset==2.2.0
django-mptt==0.16.0
django-mptt-admin==2.6.1
django-nested-inline==0.4.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Markdown==3.5.2
packaging==23.2
pluggy==1.4.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytest==8.0.2
pytest-cov==4.1.0
pytest-django==4.8.0
python-monkey-business==1.0.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
rpds-py==0.17.1
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
